+++ scripts/workflows/fstests/run_kernel_ci.sh
@@ -4,6 +4,41 @@
 source ${TOPDIR}/.config
 source ${TOPDIR}/scripts/lib.sh
 
+TARGET_WORFKLOW="$(basename $(dirname $0))"
+TARGET_WORFKLOW_DIR="$TARGET_WORFKLOW"
+grep -q demos $0
+if [[ $? -eq 0 ]]; then
+	TARGET_WORFKLOW_DIR="demos/$TARGET_WORFKLOW"
+fi
+TARGET_WORFKLOW_NAME="$TARGET_WORFKLOW"
+STARTED_FILE=""
+ENABLE_WATCHDOG="n"
+WATCHDOG_SLEEP_TIME=100
+WATCHDOG_SCRIPT="./scripts/workflows/${TARGET_WORFKLOW_DIR}/${TARGET_WORFKLOW}_watchdog.py"
+WATCHDOG_KILL_TASKS_ON_HANG="n"
+WATCHDOG_RESET_HUNG_SYSTEMS="n"
+
+if [[ "$TARGET_WORFKLOW" == "fstests" ]]; then
+	FSTYPE="$CONFIG_FSTESTS_FSTYP"
+	TARGET_WORFKLOW_NAME="$TARGET_WORFKLOW on $FSTYPE"
+	STARTED_FILE=$FSTESTS_STARTED_FILE
+	ENABLE_WATCHDOG="$CONFIG_FSTESTS_WATCHDOG"
+	WATCHDOG_SLEEP_TIME=$CONFIG_FSTESTS_WATCHDOG_CHECK_TIME
+	WATCHDOG_KILL_TASKS_ON_HANG="$CONFIG_FSTESTS_WATCHDOG_KILL_TASKS_ON_HANG"
+	WATCHDOG_RESET_HUNG_SYSTEMS="$CONFIG_FSTESTS_WATCHDOG_RESET_HUNG_SYSTEMS"
+elif [[ "$TARGET_WORFKLOW" == "blktests" ]]; then
+	STARTED_FILE=$BLKTESTS_STARTED_FILE
+	ENABLE_WATCHDOG="$CONFIG_BLKTESTS_WATCHDOG"
+	WATCHDOG_SLEEP_TIME=$CONFIG_BLKTESTS_WATCHDOG_CHECK_TIME
+	WATCHDOG_KILL_TASKS_ON_HANG="$CONFIG_BLKTESTS_WATCHDOG_KILL_TASKS_ON_HANG"
+	WATCHDOG_RESET_HUNG_SYSTEMS="$CONFIG_BLKTESTS_WATCHDOG_RESET_HUNG_SYSTEMS"
+elif [[ "$TARGET_WORFKLOW" == "reboot-limit" ]]; then
+	STARTED_FILE=$REBOOT_LIMIT_STARTED_FILE
+	ENABLE_WATCHDOG="$CONFIG_REBOOT_LIMIT_WATCHDOG"
+	WATCHDOG_SLEEP_TIME=$CONFIG_REBOOT_LIMIT_WATCHDOG_CHECK_TIME
+	# reboot-limit has no custom watchog script
+fi
+
 kernel_ci_subject_topic()
 {
 	if [[ "$CONFIG_KERNEL_CI_ADD_CUSTOM_SUBJECT_TOPIC" != "y" ]]; then
@@ -15,12 +50,11 @@ kernel_ci_subject_topic()
 	fi
 }
 
-FSTYPE="$CONFIG_FSTESTS_FSTYP"
 RCPT="ignore@test.com"
 MAIL_FROM_MOD=""
 SSH_TARGET="ignore"
-KERNEL_CI_LOOP="${TOPDIR}/scripts/workflows/fstests/run_loop.sh"
-SUBJECT_PREFIX="$(kernel_ci_subject_topic) on $(hostname): fstests failure for $FSTYPE on test loop "
+KERNEL_CI_LOOP="${TOPDIR}/scripts/workflows/${TARGET_WORFKLOW_DIR}/run_loop.sh"
+SUBJECT_PREFIX="$(kernel_ci_subject_topic) on $(hostname): ${TARGET_WORFKLOW_NAME} failure on test loop "
 KERNEL_CI_LOOP_PID=0
 TARGET_HOSTS=$1
 
@@ -64,7 +98,7 @@ kernel_ci_post_process()
 			exit 1
 		elif [[ -f $KERNEL_CI_OK_FILE ]]; then
 			LOOP_COUNT=$(cat $KERNEL_CI_OK_FILE)
-			SUBJECT="$(kernel_ci_subject_topic) fstests $FSTYPE never failed after $LOOP_COUNT test loops"
+			SUBJECT="$(kernel_ci_subject_topic) ${TARGET_WORFKLOW_NAME} never failed after $LOOP_COUNT test loops"
 			echo $SUBJECT
 			if [[ ! -s $KERNEL_CI_LOGTIME_FULL ]]; then
 				echo "Full run time of kernel-ci loop:"
@@ -87,14 +121,14 @@ kernel_ci_post_process()
 		exit 1
 	elif [[ -f $KERNEL_CI_OK_FILE ]]; then
 		LOOP_COUNT=$(cat $KERNEL_CI_OK_FILE)
-		SUBJECT="$(kernel_ci_subject_topic): fstests on $FSTYPE achieved steady-state goal of $LOOP_COUNT test loops!"
+		SUBJECT="$(kernel_ci_subject_topic): ${TARGET_WORFKLOW_NAME} achieved steady-state goal of $LOOP_COUNT test loops!"
 		GOAL="$CONFIG_KERNEL_CI_STEADY_STATE_GOAL"
 		if [[ "$CONFIG_KERNEL_CI_ENABLE_STEADY_STATE" == "y" &&
 		      "$LOOP_COUNT" -lt "$CONFIG_KERNEL_CI_STEADY_STATE_GOAL" ]]; then
-			SUBJECT="$(kernel_ci_subject_topic): fstests on $FSTYPE bailed out on loop $LOOP_COUNT before steady-state goal of $GOAL!"
+			SUBJECT="$(kernel_ci_subject_topic): ${TARGET_WORFKLOW_NAME} bailed out on loop $LOOP_COUNT before steady-state goal of $GOAL!"
                 fi
 		if [[  -f $KERNEL_CI_WATCHDOG_FAIL_LOG ]]; then
-			SUBJECT="$(kernel_ci_subject_topic): fstests on $FSTYPE detected a hang after $LOOP_COUNT test loops"
+			SUBJECT="$(kernel_ci_subject_topic): ${TARGET_WORFKLOW_NAME} detected a hang after $LOOP_COUNT test loops"
 		fi
 
 		cat $KERNEL_CI_FAIL_LOG | mail -s "'$SUBJECT'" $MAIL_FROM_MOD $RCPT
@@ -106,7 +140,7 @@ kernel_ci_post_process()
 			exit 0
 		fi
 	elif [[  -f $KERNEL_CI_WATCHDOG_FAIL_LOG ]]; then
-		SUBJECT="$(kernel_ci_subject_topic): fstests $FSTYPE failed on a hung test on the first loop"
+		SUBJECT="$(kernel_ci_subject_topic): ${TARGET_WORFKLOW_NAME} failed on a hung test on the first loop"
 		cat $KERNEL_CI_WATCHDOG_FAIL_LOG
 		cat $KERNEL_CI_DIFF_LOG | mail -s "'$SUBJECT'" $MAIL_FROM_MOD $RCPT
 		exit 1
@@ -116,7 +150,7 @@ kernel_ci_post_process()
 		echo "created if no failure was found. We did not find either file."
 		echo "This is an unexpected situation."
 
-		SUBJECT="$(kernel_ci_subject_topic): fstests $FSTYPE exited in an unexpection situation"
+		SUBJECT="$(kernel_ci_subject_topic): ${TARGET_WORFKLOW_NAME} exited in an unexpection situation"
 		cat $KERNEL_CI_LOGTIME_FULL | mail -s "'$SUBJECT'" $MAIL_FROM_MOD $RCPT
 		exit 1
 	fi
@@ -125,7 +159,6 @@ kernel_ci_post_process()
 kernel_ci_watchdog_loop()
 {
 	echo starting watchdog loop >> $KERNEL_CI_WATCHDOG_LOG
-	WATCHDOG_SLEEP_TIME=$CONFIG_FSTESTS_WATCHDOG_CHECK_TIME
 	while true; do
 		if [[ -f $MANUAL_KILL_NOTICE_FILE ]]; then
 			exit 1
@@ -135,7 +168,7 @@ kernel_ci_watchdog_loop()
 		echo watchdog loop work >> $KERNEL_CI_WATCHDOG_LOG
 		rm -f $KERNEL_CI_WATCHDOG_FAIL_LOG $KERNEL_CI_WATCHDOG_HUNG $KERNEL_CI_WATCHDOG_TIMEOUT
 
-		if [[ ! -f $FSTESTS_STARTED_FILE ]]; then
+		if [[ ! -f $STARTED_FILE ]]; then
 			if [[ ! -d /proc/$KERNEL_CI_LOOP_PID ]]; then
 				# If $KERNEL_CI_FAIL_FILE doesn't exist and we have the $KERNEL_CI_OK_FILE file
 				# we've reached steady state, so don't send extra information to stdout to
@@ -145,7 +178,7 @@ kernel_ci_watchdog_loop()
 				fi
 				break
 			fi
-			echo watchdog does not yet see start file $BLKTESTS_STARTED_FILE so waiting $WATCHDOG_SLEEP_TIME seconds >> $KERNEL_CI_WATCHDOG_LOG
+			echo watchdog does not yet see start file $STARTED_FILE so waiting $WATCHDOG_SLEEP_TIME seconds >> $KERNEL_CI_WATCHDOG_LOG
 			sleep $WATCHDOG_SLEEP_TIME
 			continue
 		fi
@@ -156,8 +189,13 @@ kernel_ci_watchdog_loop()
 			break
 		fi
 
-		echo calling blktests_watchdog.py to output into $KERNEL_CI_WATCHDOG_RESULTS_NEW >> $KERNEL_CI_WATCHDOG_LOG
-		./scripts/workflows/fstests/fstests_watchdog.py ./hosts $TARGET_HOSTS > $KERNEL_CI_WATCHDOG_RESULTS_NEW
+		if [[ ! -f $WATCHDOG_SCRIPT ]]; then
+			sleep $WATCHDOG_SLEEP_TIME
+			continue
+		fi
+
+		echo calling $(basename $WATCHDOG_SCRIPT) to output into $KERNEL_CI_WATCHDOG_RESULTS_NEW >> $KERNEL_CI_WATCHDOG_LOG
+		$WATCHDOG_SCRIPT ./hosts $TARGET_HOSTS > $KERNEL_CI_WATCHDOG_RESULTS_NEW
 		# Use the KERNEL_CI_WATCHDOG_RESULTS file to get fast results
 		cp $KERNEL_CI_WATCHDOG_RESULTS_NEW $KERNEL_CI_WATCHDOG_RESULTS
 
@@ -175,10 +213,10 @@ kernel_ci_watchdog_loop()
 			rm -f $KERNEL_CI_WATCHDOG_TIMEOUT
 		fi
 
-		if [[ "$CONFIG_FSTESTS_WATCHDOG_KILL_TASKS_ON_HANG" == "y" ]]; then
+		if [[ "$WATCHDOG_KILL_TASKS_ON_HANG" == "y" ]]; then
 			if [[ "$HUNG_FOUND" == "True" || "$TIMEOUT_FOUND" == "True" ]]; then
-				${TOPDIR}/scripts/workflows/fstests/kill_pids.sh --watchdog-mode 2> /dev/null
-				echo "The kdevops fstests watchdog detected hung or timed out hosts, stopping" >> $KERNEL_CI_WATCHDOG_FAIL_LOG
+				${TOPDIR}/scripts/workflows/${TARGET_WORFKLOW_DIR}/kill_pids.sh --watchdog-mode 2> /dev/null
+				echo "The kdevops ${TARGET_WORFKLOW} watchdog detected hung or timed out hosts, stopping" >> $KERNEL_CI_WATCHDOG_FAIL_LOG
 				echo "all tests as otherwise we'd never have this test complete, so we killed PID $KERNEL_CI_LOOP_PID." >> $KERNEL_CI_WATCHDOG_FAIL_LOG
 				echo "" >> $KERNEL_CI_WATCHDOG_FAIL_LOG
 				echo "These are critical issues, you should try to reproduce manually and fix them." >> $KERNEL_CI_WATCHDOG_FAIL_LOG
@@ -193,7 +231,7 @@ kernel_ci_watchdog_loop()
 					grep runtime $KERNEL_CI_WATCHDOG_RESULTS >> $KERNEL_CI_WATCHDOG_FAIL_LOG
 					cat $KERNEL_CI_WATCHDOG_TIMEOUT >> $KERNEL_CI_WATCHDOG_FAIL_LOG
 				fi
-				if [[ "$CONFIG_FSTESTS_WATCHDOG_RESET_HUNG_SYSTEMS" == "y" ]]; then
+				if [[ "$WATCHDOG_RESET_HUNG_SYSTEMS" == "y" ]]; then
 					for i in $(awk '{print $1}' $KERNEL_CI_WATCHDOG_RESULTS | egrep -v "runtime|Hostname"); do
 						sudo virsh reset vagrant_$i
 						echo -e "\nReset all your associated systems:" >> $KERNEL_CI_WATCHDOG_FAIL_LOG
@@ -209,7 +247,7 @@ kernel_ci_watchdog_loop()
 }
 
 rm -f ${TOPDIR}/.kernel-ci.*
-rm -f $FSTESTS_STARTED_FILE
+rm -f $STARTED_FILE
 
 if [[ "$CONFIG_KERNEL_CI_EMAIL_REPORT" == "y" ]]; then
 	RCPT="$CONFIG_KERNEL_CI_EMAIL_RCPT"
@@ -219,14 +257,14 @@ if [[ "$CONFIG_KERNEL_CI_EMAIL_MODIFY_FROM" == "y" ]]; then
 	MAIL_FROM_MOD="-S from='$CONFIG_KERNEL_CI_EMAIL_FROM'"
 fi
 
-if [[ "$CONFIG_FSTESTS_WATCHDOG" == "y" ]]; then
+if [[ "$ENABLE_WATCHDOG" == "y" ]]; then
 	rm -f $KERNEL_CI_WATCHDOG_RESULTS_NEW $KERNEL_CI_WATCHDOG_RESULTS
 fi
 
 /usr/bin/time -f %E -o $KERNEL_CI_LOGTIME_FULL $KERNEL_CI_LOOP &
 KERNEL_CI_LOOP_PID=$!
 
-if [[ "$CONFIG_FSTESTS_WATCHDOG" != "y" ]]; then
+if [[ "$ENABLE_WATCHDOG" != "y" ]]; then
 	echo Skipping watchdog and just waiting for kernel-ci PID to complete >> $KERNEL_CI_WATCHDOG_LOG
 	wait
 else
